pyproject.toml
[tool.poetry]
name = "deepvelo"
version = "1.0.0"
description = "Deep Velocity"
authors = ["subercui <subercui@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
scvelo = "^0.2.4"
torch = "^1.2"
umap-learn = "^0.5.2"
seaborn = "^0.11.2"
adjustText = "^0.7.3"
scanpy = "^1.8.2"
numpy = "^1.21.1"
tqdm = "^4.62.3"
ipywidgets = "^7.6.5"
dgl-cu101 = "^0.4"
igraph = "^0.9.10"
hnswlib = "^0.6.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.7.0"
ipdb = "^0.13.9"
cellrank = "1.5.0"
mpi4py = "^3.1.3"
petsc = "^3.17.0"
petsc4py = "^3.17.0"
slepc = "^3.17.1"
slepc4py = "^3.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
